refactor(runtime): move SlotList to a sibling file
Resolving #9001 against main pushed shapes.rs to 2036 lines, over the
repo's 2000-line cap. Move `SlotList` (the per-hash slot list this PR
adds) and `record_shape_scan_outcome` into a sibling module; both are
self-contained, so nothing else changes.
